% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanslist_api.R
\name{deanslist_api}
\alias{deanslist_api}
\title{R API for DeansList's RESTful API}
\usage{
deanslist_api(endpoint, domain, key, endpoint_version = "v1", ...)
}
\arguments{
\item{endpoint}{the endpoint from which to requrest data.}

\item{domain}{your DearnsList domain.}

\item{key}{your school specific API key.}

\item{endpoint_version}{the endpoint verrsion.  Default is "v1".  Anything else implements "beta"
endpoints, for which you should supply the table name (i.e., supply "behavavior" to hit
the \code{get-behavior-data.php} endpoint).}

\item{...}{optional DeansList paramaters added to query string in the GET request.}
}
\value{
a deanslistr object, which is a list contain the response, endpoint path,
and returned content
}
\description{
R API for DeansList's RESTful API
}
\details{
This is a workhourse function that wraps the HTTP RESTful API that DeansList has implemented.
You will need your school's/district's/network's domain (i.e., DOMAIN in \code{https://DOMAIN.deanslistsoftware.com})
and each schools API key, which you'll need to request from DeansList.

For requests in the v1 API you give the end of the endpoint name to he \code{deanslist_api()}. For example,
the \code{/api/v1/suspensions} endpoint is accessed with
\code{deanslist_api('suspensions', domain = 'dlacademy', key = 'SUPER_SECRET_KEY')}.  Requests for the beta API
simply require the component of the end point that unique identifies (i.e., the table name).  So the \code{get-behavior-data.php} endpoint
is accessed with \code{deanslist_api('behaviors', domain = 'dlacademy', key = 'SUPER_SECRET_KEY', endpoint_version = 'beta').}
}
\examples{
\dontrun{
dla_key <- Sys.getenv("KEY_DLA")

x <- deanslist_api(endpoint = 'suspensions',
                   domain = 'dlacademy',
                   key = 'dla_key')
}

}

