% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanslistr.R
\docType{package}
\name{deanslistr}
\alias{deanslistr}
\alias{deanslistr-package}
\title{deanslistr}
\description{
deanslistr is an R API for the DeansList RESTful API
}
\section{Main functions}{

\code{deanslist_api()}
}

\section{Package context}{

For getting data out of DeansList for analysis as well as
transformation and loading into your various data stores.
}
\keyword{internal}

