% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl.R
\name{get_}
\alias{get_}
\title{get_* consturctor function}
\usage{
get_(endpoint, key_list, ...)
}
\arguments{
\item{endpoint}{name of DeansList endpoint to construct function for}

\item{...}{parameteres like domain passed on onto \code{\link{deanslist_api}}}

\item{keylist}{either a named list or named character vector of DeansLists API keys for each school
to accdess suspesions data for your school(s).  Alternatively you can add keys to .Renviron variables, which must start with \code{DL_KEY_}
in order to be recognized.}
}
\description{
get_* consturctor function
}

