% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl.R
\name{get_row_count}
\alias{get_row_count}
\title{get's rowcount from returned deanslist objects}
\usage{
get_row_count(x)
}
\arguments{
\item{x}{object with content slot returned from deanslist}
}
\description{
get's rowcount from returned deanslist objects
}

