% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl.R
\name{get_students}
\alias{get_students}
\title{Wrapper around \code{\link{deanslist_api}} to pull student data for multiple schools}
\usage{
get_students(key_list, ...)
}
\arguments{
\item{key_list}{either a named list or named character vector of DeansLists API keys for each school
to accdess suspesions data for your school(s).  Alternatively you can add keys to .Renviron variables, which must start with \code{DL_KEY_}
in order to be recognized.}

\item{...}{parameteres like domain passed on onto \code{\link{deanslist_api}}}
}
\value{
a data.frame or list of data.frames (if multiple keys passed) with DeansList student data
}
\description{
This function wraps \code{\link{deanslist_api}} to pull the suspenions endpoint
(i.e., \code{deanlist_api(endpoint = 'suspensions', ...)}.  If you don't explicitly pass the
\code{key_list} paramater a named list, the function will attempt use \code{\link[base]{Sys.getenv}}
to gather keys from environment variables.  These variables can be added to a \code{.Renviron} file
that is source on startup or you can you set them using \code{\link[base]{Sys.setenv}}.  In either case,
the function looks for variables that begin with \code{DLAKEY_}.  So for Langley High School you might
set a key with the variable name \code{DL_KEY_LHS = 'super_secret_key'}.

If multiple keys are found (or passed as a named list to the \code{key_list} argument), then
a list of data frames will be returned.  The slot in the list for each data frame is named
after the key used to retrieve it.
}

