% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl.R
\name{get_suspensions}
\alias{get_suspensions}
\title{Wrapper around \code{\link{deanslist_api}} to pull suspesion data for multiple schools}
\usage{
get_suspensions(key_list, ...)
}
\arguments{
\item{key_list}{either a names list of or character vector of DeansLists API keys for each school
to accdess suspesions data for or you can set .Renviron variables}

\item{...}{parameteres like domain passed on onto \code{\link{deanslist_api}}}
}
\value{
a data.frame with DeansList suspension data
}
\description{
Wrapper around \code{\link{deanslist_api}} to pull suspesion data for multiple schools
}

