% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanslistr.R
\docType{package}
\name{deanslistr}
\alias{deanslistr}
\alias{deanslistr-package}
\title{deanslistr}
\description{
deanslistr is an R API for the DeansList RESTful API
}
\section{Main functions}{

\code{deanslist_api()}
}

\section{Package context}{

}
\keyword{internal}

